split scid checks into separate file
